decouple Event manager
